Merged PR 933: Added dockerfiles and reduced the dependency footprint of the project
I have added a `dockerfile` and `dockerignore` file, which can be used to build this project using docker.

I've also reduced the dependency footprint of this project. It no longer requires the entirety for `Microsoft.AspNetCore` to be restored before it can be built, as it only actually requires `Microsoft.AspNetCore.Http.Abstractions` in order to build the project